fix: getting the latest artifact from the correct ci file
We changed the way the ci in ArkScript/Ark worked, so we had to change this here to since we are getting artifacts from Ark (to avoid building it again) to run the standard library tests.

The file in which the ci is defined changed, as well as the artifact names. Here we are just taking one made for Ubuntu, the others should work as well (this is already being tested in the Ark ci on each platform and compiler when a PR / new commit to dev is made).